Removed and renamed communications modules
