--- Modules/Communications/config.py
@@ -1,4 +1,0 @@
-HTTP_HOST = ""
-HTTP_PORT = 8080
-
-LIVE_DISPLAY_FILE_NAME = 'Modules/Communications/html/live.html'
