--- Modules/Communications/html/live.html
@@ -1,157 +1,0 @@
-<html>
-
-<head>
-    <script type='text/javascript'>
-        String.prototype.alert = function() {
-            document.getElementById("error").innerText = this;
-        }
-        //function to set torquer values
-        function updateTorquer(dict) {
-            var str = "";
-            var keys = Object.keys(dict);
-            str += keys[0] + "=" + dict[keys[0]];
-            for (i = 1; i < keys.length; i++) {
-                str += "&";
-                str += keys[i] + "=" + dict[keys[i]];
-            }
-            console.log(str);
-            //making the http request to set 
-            httpRequest("set_magnetorquer?" + str).then((raw) => {;
-            });
-        }
-        //function that make assync http request
-        function httpRequest(url) {
-            return new Promise((success, error) => {
-                var xmlhttp = new XMLHttpRequest();
-                xmlhttp.onreadystatechange = function() {
-                    if (this.readyState == 4) {
-                        if (this.status == 200) {
-                            success(xmlhttp.responseText);
-                        } else {
-                            error(xmlhttp.status);
-                        }
-                    }
-                };
-                xmlhttp.open("GET", url, true);
-                xmlhttp.send(null);
-            });
-        }
-        //function to parse data from server to array
-        function parse(raw) {
-            var temp = [];
-            var arr = raw.split(",");
-            for (var i = 0; i < arr.length; i++) {
-                var arr1 = arr[i].split(":");
-                temp[temp.length] = arr1;
-            }
-            return temp;
-        }
-        //the function that loops all the time and is responsible for updating magnetometer values
-        function loop() {
-            var val = "";
-            httpRequest("get_magnetometer").then((raw) => {;
-                raw = raw.replace("{", "");
-                raw = raw.replace("}", "");
-                var arr = parse(raw);
-                for (i = 0; i < arr.length; i++) {
-                    val = arr[i][1];
-                    val = parseFloat(val).toFixed(2).toString();
-                    document.getElementById("element" + i.toString()).value = val;
-                }
-                /*arr.forEach((element) => {
-                    val += element[1];
-                    document.getElementById("element"+)
-                });
-                document.getElementById("element").innerHTML = val;*/
-            });
-
-        }
-        //function that sets text box's values to magnetorquer read
-        function getToruqer() {
-            var val = "";
-            httpRequest("get_magnetorquer").then((raw) => {;
-                var arr = parse(raw);
-                console.log(arr);
-                var cords = "XYZ";
-                for (i = 0; i < arr.length; i++) {
-                    val = arr[i][1];
-                    document.getElementById("set" + cords[i]).value = val;
-                }
-            });
-        }
-        //function that sets magnetorquer values in server
-        function set() {
-            var dict = [];
-            var str = ['x', 'y', 'z'];
-            var valid = true;
-            str.forEach(function(element) {
-                dict[element] = document.getElementById("set" + element.toUpperCase()).value;
-                var value = parseFloat(dict[element]);
-                if (value != 0 && value != 1.0 && value != -1.0) {
-                    "Value can only be 1, 0 or -1".alert();
-                    valid = false;
-                }
-            });
-            if (valid) {
-                "".alert();
-                updateTorquer(dict);
-            }
-        }
-        getToruqer();
-        setInterval(loop, 1000);
-    </script>
-
-</head>
-
-<body>
-    <center>
-
-        <table style='width:50%'>
-            <tr>
-                <td align='center'>
-                    <tr>
-                        <td>
-                            <table>
-                                <tr>
-                                    <td>X:</td>
-                                    <td><input type='text' id='setX'></td>
-                                </tr>
-                                <tr>
-                                    <td>Y:</td>
-                                    <td><input type='text' id='setY'></td>
-                                </tr>
-                                <tr>
-                                    <td>Z:</td>
-                                    <td><input type='text' id='setZ'></td>
-                                </tr>
-                                <tr>
-                                    <td><input type='button' onclick='set()' style='width:100%' value='set'></td>
-                                    <td><input type='button' onclick="getToruqer()" style='width:100%' value='get'></td>
-                                </tr>
-                            </table>
-                        </td>
-
-                        <td align='center'>
-                            <table>
-                                <tr>
-                                    <td>X:</td>
-                                    <td><input type='text' id='element0'></td>
-                                </tr>
-                                <tr>
-                                    <td>Y:</td>
-                                    <td><input type='text' id='element1'></td>
-                                </tr>
-                                <tr>
-                                    <td>Z:</td>
-                                    <td><input type='text' id='element2'></td>
-                                </tr>
-                            </table>
-                        </td>
-                    </tr>
-        </table>
-        <h4 style='color:red' id='error'></h4>
-        <div id="element"></div>
-    </center>
-</body>
-
-</html>
